Adjust height or speed of fish
Added functionality where the user can use the up/down arrow keys to adjust the height on the page of the fish's sinusoidal path within certain bounds. The user can also use the left and right arrows to slow down and speed up the fish, respectively. But as of this version, the change in speed will only take place once the fish changes direction and therefore is not immediately visible.